refactor flow decls
