+++ flow-typed/decls/todosForm.js
@@ -1,0 +1,13 @@
+declare type TodosFormModelSchema = {
+  input: string
+}
+
+declare type TodosFormViewModelSchema = {}
+
+declare class TodosFormModel {
+  getInput: () => string,
+  updateInput: (value: string) => TodosModel
+}
+
+declare class TodosFormViewModel extends TodosFormModel {
+}
